Finished borrarprocs
The SHELL_EXIT_SIGNAL is a definition I made back when we started P1, and it can be returned by any command to tell main( ) to clean up and exit. It should not be zero because that's the EXIT_SUCESS definition. I probably changed it by mistake, don't even remember...